% Generated by roxygen2 (4.0.1): do not edit by hand
\name{c_score_skew}
\alias{c_score_skew}
\title{Calculate the skew of the c-score}
\usage{
c_score_skew(m = matrix(rbinom(100, 1, 0.5), nrow = 10))
}
\description{
C-score skew function Takes a binary presence absence matrix returns skewness of Stone and Roberts C-score
}

