% Generated by roxygen2 (4.0.2): do not edit by hand
\name{c_score_var}
\alias{c_score_var}
\title{Calculate variance of the c-score}
\usage{
c_score_var(m = matrix(rbinom(100, 1, 0.5), nrow = 10))
}
\description{
C-score variance function Takes a binary presence absence matrix returns variance of Stone and Roberts C-score
}

