% Generated by roxygen2 (4.0.1): do not edit by hand
\name{czekanowski_skew}
\alias{czekanowski_skew}
\title{Czekanowski skew function}
\usage{
czekanowski_skew(m = matrix(rpois(80, 1), nrow = 10))
}
\description{
Takes a niche utilization matrix returns skew of the Czekanowski niche overlap index
}

