% Generated by roxygen2 (4.0.1): do not edit by hand
\name{czekanowski_var}
\alias{czekanowski_var}
\title{Czekanowski variance function}
\usage{
czekanowski_var(m = matrix(rpois(80, 1), nrow = 10))
}
\description{
Takes a niche utilization matrix returns variance Czekanowski niche overlap index
}

