% Generated by roxygen2 (4.0.2): do not edit by hand
\name{niche_null_model}
\alias{niche_null_model}
\title{Niche overlap}
\usage{
niche_null_model(species_data, algo = "ra3", metric = "Pianka",
  n.reps = 1000, row.names = TRUE, random.seed = 0)
}
\arguments{
\item{species_data}{a dataframe <put some guidelines in here>}

\item{algo}{the algorithm to use, must be "ra1", "ra2", "ra3", "ra4"}

\item{metric}{the metric used to caluclate the null model: choices are "Pianka", "Czekanowski", "Pianka.var", "Czekanowski.var", "Pianka.skew", "Czekanowski.skew"; default is Pianka}

\item{n.reps}{the number of replicates to run the null model.}

\item{row.names}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{random.seed}{Choose a seed to start your random number.  0 will choose a random seed, otherwise set the seed with any integer.}
}
\description{
Create a niche overlap null model
}
\examples{
\dontrun{
## Load MacAruthur warbler data
data(macwarb)

## Run the null model
warbMod <- niche_null_model(macwarb)
## Summary and plot info
summary(warbMod)


}
}

