% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null_model_engine.R
\name{null_model_engine}
\alias{null_model_engine}
\title{Run null model}
\usage{
null_model_engine(speciesData, algo, metric, nReps = 1000, rowNames = TRUE,
  saveSeed = FALSE, algoOpts = list(), metricOpts = list())
}
\arguments{
\item{speciesData}{a dataframe <put some guidelines in here>}

\item{algo}{the algorithm to use, must be "RA1", "RA2", "RA3", "RA4"}

\item{metric}{the metric used to caluclate the null model: choices are "Pianka", "Czekanowski", "Pianka.var", "Czekanowski.var", "Pianka.skew", "Czekanowski.skew"; default is Pianka}

\item{nReps}{the number of replicates to run the null model.}

\item{rowNames}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{saveSeed}{Should the existing random seed be saved to make the model reproducible?}

\item{algoOpts}{a list containing options for a supplied alogrithm}

\item{metricOpts}{a list containing options for a supplied metric}
}
\description{
This drives the null models for all the different kinds of null models that can be run. It is the underlying engine.
}

