% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coccurrence_null.R
\name{plot.coocnullmod}
\alias{plot.coocnullmod}
\title{Co-Occurrence Model Plot function}
\usage{
\method{plot}{coocnullmod}(x, type = "hist", ...)
}
\arguments{
\item{x}{the null model to plot}

\item{type}{the type of null plot to make.  See details for more information}

\item{...}{Other variables to be passed on to base plotting}
}
\description{
Plot co-occurrence null model object.
}
\details{
the valid types for size are "hist" to show a histogram and "size" to show a sample size null model.
}

