% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/niche_overlap_null.R
\name{plot.nichenullmod}
\alias{plot.nichenullmod}
\title{Niche Null Model Plot function}
\usage{
\method{plot}{nichenullmod}(x, type = "hist", ...)
}
\arguments{
\item{x}{the null model to plot}

\item{type}{the type of null plot to make.  See details for more information}

\item{...}{Other variables to be passed on to base plotting}
}
\description{
Generic function for plotting a histogram of simulated values Takes as input a list of Null.Model.Out, with Obs and Sim values
}
\details{
the valid types for size are "hist" to show a histogram and "niche" to show a sample draw from the null model.
}

