% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null_model_engine.R
\name{plot.nullmod}
\alias{plot.nullmod}
\title{plot a histogram null model}
\usage{
\method{plot}{nullmod}(x, ...)
}
\arguments{
\item{x}{the null model to plot}

\item{...}{Other variables to be passed on to base plotting}
}
\description{
plot a histogram of a generic null model object
}
\details{
the valid types for size are "hist" to show a histogram and "size" to show a sample size null model.
}

