% Generated by roxygen2 (4.1.0): do not edit by hand
\name{plot.sizenullmod}
\alias{plot.sizenullmod}
\title{plot a size null model}
\usage{
\method{plot}{sizenullmod}(x, type = "hist", ...)
}
\arguments{
\item{x}{the null model to plot}

\item{type}{the type of null plot to make.  See details for more information}

\item{...}{Other variables to be passed on to base plotting}
}
\description{
plot a variety of size null models
}
\details{
the valid types for size are "hist" to show a histogram and "size" to show a sample size null model.
}

