% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim10}
\alias{sim10}
\title{SIM9 function Takes a binary presence absence matrix returns a new matrix with same number of rows and columns uses swapping function from Denitz
Sim10 Function}
\usage{
sim10(speciesData, rowWeights, colWeights)
}
\description{
Randomizes a binary matrix speciesData by reshuffling all elements Rows & columns proportional to supplied row and column weights Makes a call to the VectorSample
}

