% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim4}
\alias{sim4}
\title{Sim4 Function}
\usage{
sim4(speciesData)
}
\description{
Randomizes a binary matrix speciesData by reshuffling elements within each row Sampling weights for each column are proportional to column sums Makes a call to the VectorSample
}

