% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim5}
\alias{sim5}
\title{Sim5 Function}
\usage{
sim5(speciesData)
}
\description{
Randomizes a matrix speciesData by reshuffling elements within each columnSampling weights for each row are proportional to row sums Makes a call to the VectorSample.
}

