\name{Sim6}
\alias{Sim6}
\title{Sim6 Function
Randomizes a binary matrix m by reshuffling all elements Rows are equiprobable, columns proportional to column sums Makes a call to the VectorSample.}
\usage{
  Sim6(m)
}
\description{
  Sim6 Function Randomizes a binary matrix m by reshuffling
  all elements Rows are equiprobable, columns proportional
  to column sums Makes a call to the VectorSample.
}

