% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim7}
\alias{sim7}
\title{Sim7 Function}
\usage{
sim7(speciesData)
}
\description{
Randomizes a binary matrix speciesData by reshuffling all elements Columns are equiprobable, rows proportional to row sums Makes a call to the VectorSample
}

