% Generated by roxygen2 (4.0.2): do not edit by hand
\name{source_pool_draw}
\alias{source_pool_draw}
\title{Source pool of body sizes}
\usage{
source_pool_draw(speciesData = 21:30, sourcePool = runif(n = 2 *
  length(speciesData), min = 10, max = 50), speciesProbs = rep(1,
  length(sourcePool)))
}
\description{
Function to randomize body sizes by drawing from a user-defined source pool. Species are drawn without replacement, and there is a specified probability vector for the source pool species
}

