% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.coocnullmod}
\alias{summary.coocnullmod}
\title{Generic function for calculating null model summary statistics}
\usage{
\method{summary}{coocnullmod}(object, ...)
}
\arguments{
\item{object}{the null model object to print a summary.}
}
\description{
Takes as input a list of Null.Model.Out, with Obs, Sim, Elapsed Time, and Time Stamp values
}

