% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/niche_overlap_null.R
\name{summary.nichenullmod}
\alias{summary.nichenullmod}
\title{Generic function for calculating null model summary statistics}
\usage{
\method{summary}{nichenullmod}(object, ...)
}
\arguments{
\item{object}{the null model object to print a summary of.}

\item{...}{Extra parameters for summary}
}
\description{
Takes as input a list of Null.Model.Out, with Obs, Sim, Elapsed Time, and Time Stamp values
}

