% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.nullmod}
\alias{summary.nullmod}
\title{Generic function for calculating null model summary statistics}
\usage{
\method{summary}{nullmod}(object, ...)
}
\arguments{
\item{object}{the null model object to print a summary of.}

\item{...}{Extra parameters for summary}
}
\description{
Takes a null model object and prints a nice summary.
}

