% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sizeratio_null.R
\name{summary.sizenullmod}
\alias{summary.sizenullmod}
\title{Generic function for calculating null model summary statistics}
\usage{
\method{summary}{sizenullmod}(object, ...)
}
\arguments{
\item{object}{the null model object to print a summary of.}

\item{...}{Extra parameters for summary}
}
\description{
Takes as input a list of Null.Model.Out, with Obs, Sim, Elapsed Time, and Time Stamp values
}

