% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cooc_null_model}
\alias{cooc_null_model}
\title{Co-Occurrence Null model}
\usage{
cooc_null_model(speciesData, algo = "simFast", metric = "C.Score",
  n.reps = 1000, rowNames = TRUE, randomSeed = 0, burnin = 0,
  algoOpts = list(), metricOpts = list())
}
\arguments{
\item{speciesData}{a dataframe <put some guidelines in here>}

\item{algo}{the algorithm to use, must be "sim1", "sim2", "sim3", "sim4", "sim5", "sim6", "sim7", "sim8", "sim9", "sim10"}

\item{metric}{the metric used to caluclate the null model: choices are "Species.Combo", "Checker", "C.Score", "C.Score.var", "C.Score.skew", "V.Ratio"; default is "C.Score"}

\item{rowNames}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{randomSeed}{Choose a seed to start your random number.  0 will choose a random seed, otherwise set the seed with any integer.}

\item{burnin}{The number of burnin iterations to use with the simFast algorithm}

\item{nReps}{the number of replicates to run the null model.}
}
\description{
Create a Co-Occurrence null model
}
\examples{
\dontrun{

## Run the null model
finchMod <- cooc_null_model(wiFinches, algo="sim1")
## Summary and plot info
summary(finchMod)
plot(finchMod,type="burnin")

}
}

