% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/niche_overlap_null.R
\name{niche_null_model}
\alias{niche_null_model}
\title{Niche overlap}
\usage{
niche_null_model(speciesData, algo = "ra3", metric = "pianka",
  nReps = 1000, rowNames = TRUE, saveSeed = FALSE, algoOpts = list(),
  metricOpts = list())
}
\arguments{
\item{speciesData}{a dataframe <put some guidelines in here>}

\item{algo}{the algorithm to use, must be "ra1", "ra2", "ra3", "ra4"}

\item{metric}{the metric used to caluclate the null model: choices are "pianka", "czekanowski", "pianka_var", "czekanowski_var", "pianka_skew", "czekanowski_skew"; default is pianka}

\item{nReps}{the number of replicates to run the null model.}

\item{rowNames}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{saveSeed}{TRUE or FALSE.  If TRUE the current seed is saved so the simulation can be repeated}

\item{algoOpts}{a list containing all the options for the specific algorithm you want to use.  Must match the algorithm given in the `algo` argument}

\item{metricOpts}{a list containing all the options for the specific metric you want to use.  Must match the metric given in the `metric` argument}
}
\description{
Create a niche overlap null model
}
\examples{
\dontrun{
## Load MacAruthur warbler data
data(dataMacWarb)

## Run the null model
warbMod <- niche_null_model(dataMacWarb,nReps=100000)
## Summary and plot info
summary(warbMod)
plot(warbMod)
plot(warbMod,type="niche")
}
}

