% Generated by roxygen2 (4.0.2): do not edit by hand
\name{niche_null_model}
\alias{niche_null_model}
\title{Niche overlap null models}
\usage{
niche_null_model(speciesData, algo = "ra3", metric = "pianka",
  nReps = 1000, rowNames = TRUE, saveSeed = FALSE, algoOpts = list(),
  metricOpts = list())
}
\arguments{
\item{speciesData}{a data frame in which each row is a species, each column is a resource utilization category, and the entries represent the quantity of the resource used by each species. Examples might be the amount of time a species spends foraging in different microhabitats, the biomass of different prey types, or counts of the number of times an adult female oviposits eggs on different species of a host plant.}

\item{algo}{the algorithm to use, must be "ra1", "ra2", "ra3", "ra4"}

\item{metric}{the metric used to caluclate the null model: choices are "pianka", "czekanowski", "pianka_var", "czekanowski_var", "pianka_skew", "czekanowski_skew"; default is pianka}

\item{nReps}{the number of replicates to run the null model.}

\item{rowNames}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{saveSeed}{TRUE or FALSE.  If TRUE the current seed is saved so the simulation can be repeated}

\item{algoOpts}{a list containing all the options for the specific algorithm you want to use.  Must match the algorithm given in the `algo` argument}

\item{metricOpts}{a list containing all the options for the specific metric you want to use.  Must match the metric given in the `metric` argument}
}
\description{
Create a niche overlap null models, choices of algorithm and metric are constrained to be valid for niche null models.
}
\examples{
\dontrun{
## Load MacAruthur warbler data
data(dataMacWarb)

## Run the null model
warbMod <- niche_null_model(dataMacWarb,nReps=100000)
## Summary and plot info
summary(warbMod)
plot(warbMod)
plot(warbMod,type="niche")
}
}

