% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/null_model_engine.R
\name{null_model_engine}
\alias{null_model_engine}
\title{Run null model}
\usage{
null_model_engine(speciesData, algo, metric, nReps = 1000, rowNames = TRUE,
  saveSeed = FALSE, algoOpts = list(), metricOpts = list(), type = NULL,
  suppressProg = FALSE)
}
\arguments{
\item{speciesData}{a dataframe of data that will work with metrics and algorithms.}

\item{algo}{the algorithm to use}

\item{metric}{the metric used to caluclate the null model}

\item{nReps}{the number of replicates to run the null model.}

\item{rowNames}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{saveSeed}{Should the existing random seed be saved to make the model reproducible?}

\item{algoOpts}{a list containing options for a supplied alogrithm}

\item{metricOpts}{a list containing options for a supplied metric}

\item{type}{The type of null model you are running,  if it's meant to integrate with existing null models the type should be "size","niche!","cooc", or leave it NULL if you it's not meant to be compatible with existing null models and a generic null model will be returned.}

\item{suppressProg}{a parameter to suppress the progress bar. Mostly this is just used for creating documentation with knitr}
}
\description{
This drives the null models for all the different kinds of null models that can be run. It is the underlying engine.
}
\examples{
\dontrun{
# User defined function


}
}

