% Generated by roxygen2 (4.0.2): do not edit by hand
\name{null_model_engine}
\alias{null_model_engine}
\title{Run null model}
\usage{
null_model_engine(speciesData, algo, metric, nReps = 1000, rowNames = TRUE,
  saveSeed = FALSE, algoOpts = list(), metricOpts = list())
}
\arguments{
\item{speciesData}{a dataframe <put some guidelines in here>}

\item{algo}{the algorithm to use, must be "RA1", "RA2", "RA3", "RA4"}

\item{metric}{the metric used to caluclate the null model: choices are "Pianka", "Czekanowski", "Pianka.var", "Czekanowski.var", "Pianka.skew", "Czekanowski.skew"; default is Pianka}

\item{nReps}{the number of replicates to run the null model.}

\item{rowNames}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{random.seed}{Choose a seed to start your random number.  0 will choose a random seed, otherwise set the seed with any integer.}
}
\description{
This drives the null models for all the different kinds of null models that can be run. It is the underlying engine.
}

