% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.nichenullmod}
\alias{plot.nichenullmod}
\title{Null Model Plot function}
\usage{
\method{plot}{nichenullmod}(x, type = "hist", ...)
}
\description{
Generic function for plotting a histogram of simulated values Takes as input a list of Null.Model.Out, with Obs and Sim values
}

