% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim10}
\alias{sim10}
\title{Sim10 Function}
\usage{
sim10(m, Row.Weights, Col.Weights)
}
\description{
Randomizes a binary matrix m by reshuffling all elements Rows & columns proportional to supplied row and column weights Makes a call to the VectorSample
}

