% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sim6}
\alias{sim6}
\title{Sim6 Function
Randomizes a binary matrix m by reshuffling all elements Rows are equiprobable, columns proportional to column sums Makes a call to the VectorSample.}
\usage{
sim6(m)
}
\description{
Sim6 Function
Randomizes a binary matrix m by reshuffling all elements Rows are equiprobable, columns proportional to column sums Makes a call to the VectorSample.
}

