% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim9}
\alias{sim9}
\title{SIM9 function Takes a binary presence absence matrix returns a new matrix with same number of rows and columns uses swapping function from Denitz}
\usage{
sim9(m = matrix(rbinom(100, 1, 0.5), nrow = 10))
}
\description{
SIM9 function Takes a binary presence absence matrix returns a new matrix with same number of rows and columns uses swapping function from Denitz
}

