% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim9fast.R
\name{sim9}
\alias{sim9}
\title{sim9.fast}
\usage{
sim9(speciesData, algo, metric, nReps = 1000, rowNames = TRUE,
  saveSeed = FALSE, burn_in = 0, suppressProg = TRUE)
}
\arguments{
\item{speciesData}{a dataframe in which rows are species, columns are sites,
and the entries indicate the absence (0) or presence (1) of a species in a
site. Empty rows and empty columns should not be included in the matrix.}

\item{algo}{the algorithm to use, must be "sim1", "sim2", "sim3", "sim4", "sim5", "sim6", "sim7", "sim8", "sim9", "sim10"}

\item{metric}{the metric used to caluclate the null model: choices are "species_combo", "checker", "c_score", "c_score_var", "c_score_skew", "v_ratio"; default is "c_score"}

\item{nReps}{the number of replicates to run the null model.}

\item{saveSeed}{TRUE or FALSE.  If TRUE the current seed is saved so the simulation can be repeated}

\item{burn_in}{The number of burn_in iterations to use with the simFast algorithm}

\item{suppressProg}{a parameter to suppress the progress bar. Mostly this is just used for creating documentation with knitr}

\item{algoOpts}{a list containing all the options for the specific algorithm you want to use.  Must match the algorithm given in the `algo` argument}

\item{metricOpts}{a list containing all the options for the specific metric you want to use.  Must match the metric given in the `metric` argument}
}
\description{
A special implementation of the sequential swap algorithm
}
\examples{
\dontrun{

## Run the null model
finchMod <- cooc_null_model(dataWiFinches, algo="sim1",nReps=1000000,burn_in = 500)
## Summary and plot info
summary(finchMod)
plot(finchMod,type="burn_in")
plot(finchMod,type="hist")
plot(finchMod,type="cooc")
}
}

