% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim9fast.R
\name{sim9_single}
\alias{sim9_single}
\title{sim9_single}
\usage{
sim9_single(speciesData = matrix(rbinom(100, 1, 0.5), nrow = 10))
}
\arguments{
\item{speciesData}{binary presence-absence matrix.}
}
\description{
Function for a single iteration of the sequential swap.
}
\details{
See details for sim9.
}

