% Generated by roxygen2 (4.0.1): do not edit by hand
\name{size_null_model}
\alias{size_null_model}
\title{Size Ratio}
\usage{
size_null_model(species_data, algo = "Uniform.Size", metric = "Var.Ratio",
  n.reps = 1000, row.names = TRUE, random.seed = 0)
}
\arguments{
\item{species_data}{a dataframe <put some guidelines in here>}

\item{algo}{the algorithm to use, must be "Uniform.Size", "Uniform.Size.User", "Source.Pool", "Gamma"}

\item{metric}{the metric used to caluclate the null model: choices are "Min.Diff", "Min.Ratio", "Var.Diff", "Var.Ratio"; default is Var.Ratio}

\item{n.reps}{the number of replicates to run the null model.}

\item{row.names}{Does your dataframe have row names? If yes, they are stripped, otherwise FALSE for data that has no row names}

\item{random.seed}{Choose a seed to start your random number.  0 will choose a random seed, otherwise set the seed with any integer.}
}
\description{
Create a size Ratio null model
}
\examples{
\dontrun{
## Run the null model
rodentMod <- size_null_model(rodents)
## Summary and plot info
summary(rodentMod)
plot(rodentMod,type="hist")
plot(rodentMod,type="size")

}
}

