% Generated by roxygen2 (4.0.1): do not edit by hand
\name{source_pool_draw}
\alias{source_pool_draw}
\title{Source pool of body sizes}
\usage{
source_pool_draw(v = 21:30, Source.Pool = runif(n = 2 * length(v), min = 10,
  max = 50), Species.Probs = rep(1, length(Source.Pool)))
}
\description{
Function to randomize body sizes by drawing from a user-defined source pool. Species are drawn without replacement, and there is a specified probability vector for the source pool species
}

