% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.nichenullmod}
\alias{summary.nichenullmod}
\title{Generic function for calculating null model summary statistics}
\usage{
\method{summary}{nichenullmod}(object, ...)
}
\description{
Takes as input a list of Null.Model.Out, with Obs, Sim, Elapsed Time, and Time Stamp values
}

