% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uniform_size_user}
\alias{uniform_size_user}
\title{User defined size limits}
\usage{
uniform_size_user(speciesData = runif(n = 20), userLow = 0.9 *
  min(speciesData), userHigh = 1.1 * max(speciesData))
}
\description{
Function to randomize uniformly body sizes within user-defined limits. Note that all n species are randomized in this algorithm
}

