% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uniform_size_user}
\alias{uniform_size_user}
\title{User defined size limits}
\usage{
uniform_size_user(v = runif(n = 20), user.low = 0.9 * min(v),
  user.high = 1.1 * max(v))
}
\description{
Function to randomize uniformly body sizes within user-defined limits. Note that all n species are randomized in this algorithm
}

